correct some things

+++ core/src/com/raiden/game/physics_controller/entities/AirPlane_3.java
@@ -10,7 +10,7 @@ import com.raiden.game.model.entities.MovingObjectModel;
  */
 public class AirPlane_3 extends ShipPhysics {
 
-    private static float MAXVELOCITY_Default = (float) (AirPlane_1.MAXVELOCITY_Default * 0.8);
+    private static float MAXVELOCITY_Default = AirPlane_1.MAXVELOCITY_Default * 0.8f;
 
     private final float density = 0.8f;
